refactored click listenrs
