+++ app/src/main/java/ru/surdasha/cats/presentation/CatsGlideModule.java
@@ -1,7 +1,6 @@
 package ru.surdasha.cats.presentation;
 
 import android.content.Context;
-import android.util.Log;
 
 import com.bumptech.glide.GlideBuilder;
 import com.bumptech.glide.annotation.GlideModule;
@@ -9,12 +8,10 @@ import com.bumptech.glide.load.engine.cache.InternalCacheDiskCacheFactory;
 import com.bumptech.glide.module.AppGlideModule;
 
 @GlideModule
-public final class TRBTLGlideModule extends AppGlideModule {
+public final class CatsGlideModule extends AppGlideModule {
     @Override
     public void applyOptions(Context context, GlideBuilder builder) {
-        int diskCacheSizeBytes = 1024 * 1024 * 50; // 100 MB
-//        int diskCacheSizeBytes = 100 * 1024; // 100 MB
-        builder.setDiskCache(new InternalCacheDiskCacheFactory(context, diskCacheSizeBytes))
-                .setLogLevel(Log.VERBOSE);
+        int diskCacheSizeBytes = 1024 * 1024 * 100; // 100 MB
+        builder.setDiskCache(new InternalCacheDiskCacheFactory(context, diskCacheSizeBytes));
     }
 }
