--- app/src/main/java/ru/surdasha/cats/presentation/GenericFileProvider.java
@@ -1,6 +1,0 @@
-package ru.surdasha.cats.presentation;
-import androidx.core.content.FileProvider;
-
-public class GenericFileProvider extends FileProvider {
-}
-
