--- app/src/main/java/ru/surdasha/cats/presentation/ui/all/MyDiffCalback.java
@@ -1,46 +1,0 @@
-package ru.surdasha.cats.presentation.ui.all;
-
-import java.util.List;
-
-import androidx.annotation.Nullable;
-import androidx.recyclerview.widget.DiffUtil;
-import ru.surdasha.cats.presentation.models.CatUI;
-
-public class MyDiffCalback extends DiffUtil.Callback{
-
-    List<CatUI> oldPersons;
-    List<CatUI> newPersons;
-
-    public MyDiffCalback(List<CatUI> newPersons, List<CatUI> oldPersons) {
-        this.newPersons = newPersons;
-        this.oldPersons = oldPersons;
-    }
-
-    @Override
-    public int getOldListSize() {
-        return oldPersons.size();
-    }
-
-    @Override
-    public int getNewListSize() {
-        return newPersons.size();
-    }
-
-    @Override
-    public boolean areItemsTheSame(int oldItemPosition, int newItemPosition) {
-        return oldPersons.get(oldItemPosition).getServerId() == newPersons.get(newItemPosition).getServerId();
-    }
-
-    @Override
-    public boolean areContentsTheSame(int oldItemPosition, int newItemPosition) {
-        return oldPersons.get(oldItemPosition).equals(newPersons.get(newItemPosition));
-    }
-
-    @Nullable
-    @Override
-    public Object getChangePayload(int oldItemPosition, int newItemPosition) {
-        //you can return particular field for changed item.
-        return super.getChangePayload(oldItemPosition, newItemPosition);
-    }
-}
-
