+++ app/src/main/java/ru/surdasha/cats/presentation/ui/favorites/FavoriteCatsAdapter.java
@@ -4,24 +4,18 @@ import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageButton;
-import android.widget.ImageView;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import androidx.annotation.NonNull;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.RecyclerView;
-import butterknife.BindView;
 import butterknife.ButterKnife;
 import ru.surdasha.cats.R;
-import ru.surdasha.cats.presentation.GlideApp;
-import ru.surdasha.cats.presentation.misc.ViewUtils;
 import ru.surdasha.cats.presentation.models.CatUI;
 
 public class FavoriteCatsAdapter extends
-        RecyclerView.Adapter<FavoriteCatsAdapter.ViewHolder> {
+        RecyclerView.Adapter<FavoriteCatsViewHolder> {
 
     private static final String TAG = FavoriteCatsAdapter.class.getSimpleName();
 
@@ -36,45 +30,15 @@ public class FavoriteCatsAdapter extends
     }
 
 
-    public static class ViewHolder extends RecyclerView.ViewHolder {
-        @BindView(R.id.ivCat)
-        ImageView ivCat;
-        @BindView(R.id.ibDelete)
-        ImageButton ibDelete;
-        @BindView(R.id.constraintLayout)
-        ConstraintLayout constraintLayout;
-
-        public ViewHolder(View itemView) {
-            super(itemView);
-            ButterKnife.bind(this, itemView);
-
-        }
-
-        public void bind(Context context,final CatUI model, final onDeleteCatListener listener) {
-            ibDelete.setOnClickListener(v -> listener.onDeleteCat(model));
-            ivCat.getLayoutParams().height = model.getScreenImageHeight();
-            showWithGlide(context, model);
-        }
-
-        private void showWithGlide(Context context, CatUI model) {
-            GlideApp.with(context)
-                    .load(model.getUrl())
-                    .placeholder(ViewUtils.createCircularImageDrawable(context))
-                    .override(model.getScreenImageWidth(), model.getScreenImageHeight())
-                    .into(ivCat);
-        }
-
-    }
-
     @Override
-    public ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+    public FavoriteCatsViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
         Context context = parent.getContext();
         LayoutInflater inflater = LayoutInflater.from(context);
 
         View view = inflater.inflate(R.layout.item_favorite_cat, parent, false);
         ButterKnife.bind(this, view);
 
-        ViewHolder viewHolder = new ViewHolder(view);
+        FavoriteCatsViewHolder viewHolder = new FavoriteCatsViewHolder(view);
 
         return viewHolder;
     }
@@ -92,7 +56,7 @@ public class FavoriteCatsAdapter extends
 
 
     @Override
-    public void onBindViewHolder(ViewHolder holder, int position) {
+    public void onBindViewHolder(FavoriteCatsViewHolder holder, int position) {
         CatUI item = list.get(position);
         holder.bind(context, item, onDeleteCatListener);
     }
