+++ app/src/main/java/ru/surdasha/cats/presentation/ui/favorites/FavoriteCatsView.java
@@ -1,7 +1,8 @@
 package ru.surdasha.cats.presentation.ui.favorites;
 
 import com.arellomobile.mvp.MvpView;
-import com.arellomobile.mvp.viewstate.strategy.AddToEndStrategy;
+import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy;
+import com.arellomobile.mvp.viewstate.strategy.OneExecutionStateStrategy;
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType;
 
 import java.util.List;
@@ -9,24 +10,24 @@ import java.util.List;
 import ru.surdasha.cats.presentation.models.CatUI;
 
 public interface FavoriteCatsView extends MvpView {
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(AddToEndSingleStrategy.class)
     void onShowCats(List<CatUI> cats);
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(AddToEndSingleStrategy.class)
     void onShowLoading();
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(OneExecutionStateStrategy.class)
     void onHideLoad();
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(OneExecutionStateStrategy.class)
     void onShowErrorLoading();
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(AddToEndSingleStrategy.class)
     void onEmptyList();
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(OneExecutionStateStrategy.class)
     void onShowErrorDeleting();
 
-    @StateStrategyType(AddToEndStrategy.class)
+    @StateStrategyType(OneExecutionStateStrategy.class)
     void onShowSuccessDeleting(CatUI catUI);
 }
