+++ app/src/main/res/values/dimens.xml
@@ -3,16 +3,6 @@
     <dimen name="activity_horizontal_margin">16dp</dimen>
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="btn_size">48dp</dimen>
-    <dimen name="btn_size">48dp</dimen>
-    <dimen name="btn_size">48dp</dimen>
-    <dimen name="btn_size">48dp</dimen>
-    <dimen name="btn_size">48dp</dimen>
-    <dimen name="btn_size">48dp</dimen>
     <dimen name="text_size">20sp</dimen>
     <dimen name="standart_margin">8dp</dimen>
-    <dimen name="text_size">20sp</dimen>
-    <dimen name="text_size">20sp</dimen>
-    <dimen name="standart_margin">8dp</dimen>
-    <dimen name="text_size">20sp</dimen>
-    <dimen name="text_size">20sp</dimen>
 </resources>
